execution/requirements.txt
python-dotenv
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas

playwright
gspread
gspread-formatting
outscraper
psutil
